Move PostLock and StatusCheck components to wp-story-editor
