+++ packages/wp-story-editor/src/components/postLock/postLockDialog.js
@@ -23,15 +23,14 @@ import {
   Button,
   BUTTON_SIZES,
   BUTTON_TYPES,
-  BUTTON_VARIANTS,
-  Text,
   THEME_CONSTANTS,
+  Text,
 } from '@web-stories-wp/design-system';
+import { Dialog } from '@web-stories-wp/story-editor';
 
 /**
  * Internal dependencies
  */
-import Dialog from '../dialog';
 import { Img } from './shared';
 
 /**
@@ -39,35 +38,53 @@ import { Img } from './shared';
  * @param {boolean} props.isOpen If open or not.
  * @param {Object} props.user Lock owner's user data as a object.
  * @param {string} props.dashboardLink Link to dashboard.
+ * @param {string} props.previewLink Preview link.
  * @param {Function} props.onClose Function when dialog is closed.
  * @return {*} Render.
  */
-function PostTakeOverDialog({ isOpen, user, dashboardLink, onClose }) {
-  const dialogTile = __(
-    'Someone else has taken over this story.',
-    'web-stories'
-  );
+function PostLockDialog({ isOpen, onClose, user, dashboardLink, previewLink }) {
+  const dialogTile = __('Story is locked', 'web-stories');
   const dialogContent = sprintf(
     /* translators: %s: user's name */
-    __('%s now has editing control of this story.', 'web-stories'),
+    __(
+      '%s is already editing this story. Do you want to take over?',
+      'web-stories'
+    ),
     user?.name
   );
 
   return (
     <Dialog
       isOpen={isOpen}
+      onClose={onClose}
       title={dialogTile}
       contentLabel={dialogTile}
-      onClose={onClose}
       actions={
-        <Button
-          type={BUTTON_TYPES.QUATERNARY}
-          size={BUTTON_SIZES.SMALL}
-          variant={BUTTON_VARIANTS.RECTANGLE}
-          href={dashboardLink}
-        >
-          {__('My Stories', 'web-stories')}
-        </Button>
+        <>
+          <Button
+            type={BUTTON_TYPES.TERTIARY}
+            size={BUTTON_SIZES.SMALL}
+            href={dashboardLink}
+          >
+            {__('My Stories', 'web-stories')}
+          </Button>
+          <Button
+            type={BUTTON_TYPES.TERTIARY}
+            size={BUTTON_SIZES.SMALL}
+            href={previewLink}
+            target="_blank"
+            rel="noreferrer"
+          >
+            {__('Preview', 'web-stories')}
+          </Button>
+          <Button
+            type={BUTTON_TYPES.TERTIARY}
+            size={BUTTON_SIZES.SMALL}
+            onClick={onClose}
+          >
+            {__('Take over', 'web-stories')}
+          </Button>
+        </>
       }
     >
       <Text size={THEME_CONSTANTS.TYPOGRAPHY.PRESET_SIZES.SMALL}>
@@ -86,11 +103,12 @@ function PostTakeOverDialog({ isOpen, user, dashboardLink, onClose }) {
   );
 }
 
-PostTakeOverDialog.propTypes = {
+PostLockDialog.propTypes = {
   isOpen: PropTypes.bool.isRequired,
   user: PropTypes.object,
   dashboardLink: PropTypes.string.isRequired,
+  previewLink: PropTypes.string,
   onClose: PropTypes.func.isRequired,
 };
 
-export default PostTakeOverDialog;
+export default PostLockDialog;
