+++ packages/wp-story-editor/src/components/statusCheck/statusCheckFailed.js
@@ -22,10 +22,7 @@ import { useCallback } from '@web-stories-wp/react';
 import { __, TranslateWithMarkup } from '@web-stories-wp/i18n';
 import { trackClick } from '@web-stories-wp/tracking';
 import { Link, Text, THEME_CONSTANTS } from '@web-stories-wp/design-system';
-/**
- * Internal dependencies
- */
-import Dialog from '../dialog';
+import { Dialog } from '@web-stories-wp/story-editor';
 
 const SUPPORT_URL = __(
   'https://wordpress.org/support/plugin/web-stories/',
